Move NopPending to event.
